+++ eyed3/id3/tag.py
@@ -3,14 +3,14 @@ import string
 import shutil
 import tempfile
 import textwrap
-from functools import partial
 from codecs import ascii_encode
+
 
 from ..utils import requireUnicode, chunkCopy, datePicker, b
 from .. import core
 from ..core import TXXX_ALBUM_TYPE, TXXX_ARTIST_ORIGIN, ALBUM_TYPE_IDS, ArtistOrigin
 from .. import Error
-from . import (ID3_ANY_VERSION, ID3_V1, ID3_V1_0, ID3_V1_1,
+from . import (ID3_ANY_VERSION, ID3_DEFAULT_VERSION, ID3_V1, ID3_V1_0, ID3_V1_1,
                ID3_V2, ID3_V2_2, ID3_V2_3, ID3_V2_4, versionToString)
 from . import DEFAULT_LANG
 from . import Genre
@@ -20,22 +20,23 @@ from .headers import TagHeader, ExtendedTagHeader
 from ..utils.log import getLogger
 log = getLogger(__name__)
 
-
-class TagException(Error):
-    pass
-
-
 ID3_V1_COMMENT_DESC = "ID3v1.x Comment"
 ID3_V1_MAX_TEXTLEN = 30
 ID3_V1_STRIP_CHARS = string.whitespace.encode("latin1") + b"\x00"
 DEFAULT_PADDING = 256
 
 
+class TagException(Error):
+    pass
+
+
 class Tag(core.Tag):
-    def __init__(self, **kwargs):
+    def __init__(self, version=ID3_DEFAULT_VERSION, **kwargs):
+        self.file_info = None
         self.header = None
         self.extended_header = None
         self.frame_set = None
+
         self._comments = None
         self._images = None
         self._lyrics = None
@@ -47,15 +48,15 @@ class Tag(core.Tag):
         self._chapters = None
         self._tocs = None
         self._popularities = None
-        self.file_info = None
 
-        self.clear()
+        self.file_info = None
+        self.clear(version=version)
         super().__init__(**kwargs)
 
-    def clear(self):
+    def clear(self, *, version=ID3_DEFAULT_VERSION):
         """Reset all tag data."""
         # ID3 tag header
-        self.header = TagHeader()
+        self.header = TagHeader(version=version)
         # Optional extended header in v2 tags.
         self.extended_header = ExtendedTagHeader()
         # Contains the tag's frames. ID3v1 fields are read and converted
@@ -72,10 +73,8 @@ class Tag(core.Tag):
         self._chapters = ChaptersAccessor(self.frame_set)
         self._tocs = TocAccessor(self.frame_set)
         self._popularities = PopularitiesAccessor(self.frame_set)
-        self.file_info = None
 
     def parse(self, fileobj, version=ID3_ANY_VERSION):
-        assert fileobj
         self.clear()
         version = version or ID3_ANY_VERSION
 
@@ -88,7 +87,7 @@ class Tag(core.Tag):
                 fileobj = open(filename, "rb")
                 close_file = True
             else:
-                raise ValueError("Invalid type: %s" % str(type(fileobj)))
+                raise ValueError(f"Invalid type: {type(fileobj)}")
 
         self.file_info = FileInfo(filename)
 
@@ -118,6 +117,7 @@ class Tag(core.Tag):
 
     def _loadV2Tag(self, fp):
         """Returns (tag_found, padding_len)"""
+        fp.seek(0)
 
         # Look for a tag and if found load it.
         if not self.header.parse(fp):
@@ -185,7 +185,7 @@ class Tag(core.Tag):
             if (len(comment) >= 2 and
                     # Python the slices (the chars), so this is really
                     # comment[2]       and        comment[-1]
-                    comment[-2:-1] == b"\x00" and comment[-1:] != b"\x00"):
+                    comment[-2:-1] == b"\x00"):
                 log.debug("Track Num found, setting version to v1.1")
                 self.version = ID3_V1_1
 
@@ -196,11 +196,11 @@ class Tag(core.Tag):
 
             # There may only have been a track #
             if comment:
-                log.debug("Comment: %s" % comment)
+                log.debug(f"Comment: {comment}")
                 self.comments.set(str(comment, v1_enc), ID3_V1_COMMENT_DESC)
 
         genre = ord(tag_data[127:128])
-        log.debug("Genre ID: %d" % genre)
+        log.debug(f"Genre ID: {genre}")
         try:
             self.genre = genre
         except ValueError as ex:
@@ -236,7 +236,7 @@ class Tag(core.Tag):
         return self.header.major_version == 2
 
     @requireUnicode(2)
-    def setTextFrame(self, fid, txt):
+    def setTextFrame(self, fid: bytes, txt: str):
         fid = b(fid, ascii_encode)
         if not fid.startswith(b"T") or fid.startswith(b"TX"):
             raise ValueError("Invalid frame-id for text frame")
@@ -246,7 +246,8 @@ class Tag(core.Tag):
         elif txt:
             self.frame_set.setTextFrame(fid, txt)
 
-    def getTextFrame(self, fid):
+    # FIXME: is returning data not a Frame.
+    def getTextFrame(self, fid: bytes):
         fid = b(fid, ascii_encode)
         if not fid.startswith(b"T") or fid.startswith(b"TX"):
             raise ValueError("Invalid frame-id for text frame")
@@ -548,6 +549,8 @@ class Tag(core.Tag):
         elif self.version == ID3_V2_4:
             self._setDate(b"TDRC", date)
         else:
+            if not isinstance(date, core.Date):
+                date = core.Date.parse(date)
             self._setDate(b"TYER", str(date.year))
             if None not in (date.month, date.day):
                 date_str = "%s%s" % (str(date.day).rjust(2, "0"),
@@ -698,7 +701,7 @@ class Tag(core.Tag):
         if f and f[0].text:
             try:
                 return Genre.parse(f[0].text, id3_std=id3_std)
-            except ValueError:
+            except ValueError:  # pragma: nocover
                 return None
         else:
             return None
@@ -709,7 +712,7 @@ class Tag(core.Tag):
         A Genre object, an acceptable (see Genre.parse) genre string,
         or an integer genre ID all will set the value. A value of None will
         remove the genre."""
-        if g is None:
+        if g in ("", None):
             if self.frame_set[frames.GENRE_FID]:
                 del self.frame_set[frames.GENRE_FID]
             return
@@ -719,14 +722,22 @@ class Tag(core.Tag):
         elif isinstance(g, int):
             g = Genre(id=g)
         elif not isinstance(g, Genre):
-            raise TypeError("Invalid genre data type: %s" % str(type(g)))
-        self.frame_set.setTextFrame(frames.GENRE_FID, str(g))
+            raise TypeError(f"Invalid genre data type: {type(g)}")
+
+        assert g
+        self.frame_set.setTextFrame(frames.GENRE_FID, f"{g.name if g.name else g.id}")
 
     # genre property
     genre = property(_getGenre, _setGenre)
-    # Non-standard genres.
-    non_std_genre = property(partial(_getGenre, id3_std=False),
-                             partial(_setGenre, id3_std=False))
+
+    def _getNonStdGenre(self):
+        return self._getGenre(id3_std=False)
+
+    def _setNonStdGenre(self, val):
+        self._setGenre(val, id3_std=False)
+
+    # non-standard genre (unparsed, unmapped) property
+    non_std_genre = property(_getNonStdGenre, _setNonStdGenre)
 
     @property
     def user_text_frames(self):
@@ -843,8 +854,25 @@ class Tag(core.Tag):
             self.frame_set[frames.TOS_FID][0].text = tos
             self.frame_set[frames.TOS_FID][0].lang = lang
         else:
-            self.frame_set[frames.TOS_FID] = frames.TermsOfUseFrame(text=tos,
-                                                                    lang=lang)
+            self.frame_set[frames.TOS_FID] = frames.TermsOfUseFrame(text=tos, lang=lang)
+
+    def _setCopyright(self, copyrt):
+        self.setTextFrame(frames.COPYRIGHT_FID, copyrt)
+
+    def _getCopyright(self):
+        if frames.COPYRIGHT_FID in self.frame_set:
+            return self.frame_set[frames.COPYRIGHT_FID][0].text
+
+    copyright = property(_getCopyright, _setCopyright)
+
+    def _setEncodedBy(self, enc):
+        self.setTextFrame(frames.ENCODED_BY_FID, enc)
+
+    def _getEncodedBy(self):
+        if frames.ENCODED_BY_FID in self.frame_set:
+            return self.frame_set[frames.ENCODED_BY_FID][0].text
+
+    encoded_by = property(_getEncodedBy, _setEncodedBy)
 
     def _raiseIfReadonly(self):
         if self.read_only:
@@ -1128,16 +1156,55 @@ class Tag(core.Tag):
         log.debug("Tag write complete. Updating FileInfo state.")
         self.file_info.tag_size = len(tag_data) + len(padding)
 
-    def _convertFrames(self, std_frames, convert_list, version):
-        """Maps frame incompatibilities between ID3 v2.3 and v2.4.
+    def _convertFrames_v1(self, std_frames, convert_list, version) -> list:
+        assert version[0] == 1
+        converted_frames = []
+
+        track_num_frame = None
+        for frame in std_frames:
+            if frame.id == frames.TRACKNUM_FID:
+                # Find track_num so it can be enforced for 1.1
+                track_num_frame = frame
+            elif frame.id == frames.COMMENT_FID and frame.description == ID3_V1_COMMENT_DESC:
+                # Comments truncated to make room for v1.1 track
+                if version == ID3_V1_1:
+                    if len(frame.text) > ID3_V1_MAX_TEXTLEN - 2:
+                        trunc_text = frame.text[:ID3_V1_MAX_TEXTLEN - 2]
+                        log.info(f"Truncating ID3 v1 comment due to tag conversion: {frame.text}")
+                        frame.text = trunc_text
+
+        # v1.1 must have a track num
+        if track_num_frame is None and version == ID3_V1_1:
+            log.info("ID3 v1.0->v1.1 conversion forces track number, defaulting to 1")
+            std_frames.append(frames.TextFrame(frames.TRACKNUM_FID, "1"))
+        # v1.0 must not
+        elif track_num_frame is not None and version == ID3_V1_0:
+            log.info("ID3 v1.1->v1.0 conversion forces deleting track number")
+            std_frames.remove(track_num_frame)
+
+        for frame in list(convert_list):
+            # Let date frames thru, the right thing will happen on save
+            if isinstance(frame, frames.DateFrame):
+                converted_frames.append(frame)
+                convert_list.remove(frame)
+
+        return converted_frames
+
+    def _convertFrames(self, std_frames, convert_list, version) -> list:
+        """Maps frame incompatibilities between ID3 tag versions.
 
         The items in ``std_frames`` need no conversion, but the list/frames
         may be edited if necessary (e.g. a converted frame replaces a frame
         in the list).  The items in ``convert_list`` are the frames to convert
         and return. The ``version`` is the target ID3 version."""
         from . import versionToString
-        from .frames import (DATE_FIDS, DEPRECATED_DATE_FIDS,
-                             DateFrame, TextFrame)
+        from .frames import DATE_FIDS, DEPRECATED_DATE_FIDS, DateFrame, TextFrame
+
+        if version[0] == 1:
+            return self._convertFrames_v1(std_frames, convert_list, version)
+
+        # Only ID3 v2.x onward
+        assert version[0] != 1
         converted_frames = []
         flist = list(convert_list)
 
@@ -1242,6 +1309,24 @@ class Tag(core.Tag):
             tsst_frame = frames.UserTextFrame(
                     description="Subtitle (converted)", text=tsst_frame.text)
             converted_frames.append(tsst_frame)
+
+        # RVAD (v2.3) --> RVA2* (2.4)
+        if version == ID3_V2_4 and b"RVAD" in [f.id for f in flist]:
+            rvad = [f for f in flist if f.id == b"RVAD"][0]
+            for rva2 in rvad.toV24():
+                converted_frames.append(rva2)
+            flist.remove(rvad)
+        # RVA2* (v2.4) --> RVAD (2.3)
+        elif version == ID3_V2_3 and b"RVA2" in [f.id for f in flist]:
+            adj = frames.RelVolAdjFrameV23.VolumeAdjustments()
+            for rva2 in [f for f in flist if f.id == b"RVA2"]:
+                adj.setChannelAdj(rva2.channel_type, rva2.adjustment * 512)
+                adj.setChannelPeak(rva2.channel_type, rva2.peak)
+                flist.remove(rva2)
+
+            rvad = frames.RelVolAdjFrameV23()
+            rvad.adjustments = adj
+            converted_frames.append(rvad)
 
         # Raise an error for frames that could not be converted.
         if len(flist) != 0:
@@ -1453,7 +1538,7 @@ class DltAccessor(AccessorBase):
                     frame.lang == (lang if isinstance(lang, bytes)
                                         else lang.encode("ascii")))
 
-        super(DltAccessor, self).__init__(fid, fs, match_func)
+        super().__init__(fid, fs, match_func)
         self.FrameClass = FrameClass
 
     @requireUnicode(1, 2)
@@ -1472,32 +1557,28 @@ class DltAccessor(AccessorBase):
 
     @requireUnicode(1)
     def remove(self, description, lang=DEFAULT_LANG):
-        return super(DltAccessor, self).remove(description,
-                                               lang=lang or DEFAULT_LANG)
+        return super().remove(description, lang=lang or DEFAULT_LANG)
 
     @requireUnicode(1)
     def get(self, description, lang=DEFAULT_LANG):
-        return super(DltAccessor, self).get(description,
-                                            lang=lang or DEFAULT_LANG)
+        return super().get(description, lang=lang or DEFAULT_LANG)
 
 
 class CommentsAccessor(DltAccessor):
     def __init__(self, fs):
-        super(CommentsAccessor, self).__init__(frames.CommentFrame,
-                                               frames.COMMENT_FID, fs)
+        super().__init__(frames.CommentFrame, frames.COMMENT_FID, fs)
 
 
 class LyricsAccessor(DltAccessor):
     def __init__(self, fs):
-        super(LyricsAccessor, self).__init__(frames.LyricsFrame,
-                                             frames.LYRICS_FID, fs)
+        super().__init__(frames.LyricsFrame, frames.LYRICS_FID, fs)
 
 
 class ImagesAccessor(AccessorBase):
     def __init__(self, fs):
         def match_func(frame, description):
             return frame.description == description
-        super(ImagesAccessor, self).__init__(frames.IMAGE_FID, fs, match_func)
+        super().__init__(frames.IMAGE_FID, fs, match_func)
 
     @requireUnicode("description")
     def set(self, type_, img_data, mime_type, description="", img_url=None):
@@ -1539,18 +1620,18 @@ class ImagesAccessor(AccessorBase):
 
     @requireUnicode(1)
     def remove(self, description):
-        return super(ImagesAccessor, self).remove(description)
+        return super().remove(description)
 
     @requireUnicode(1)
     def get(self, description):
-        return super(ImagesAccessor, self).get(description)
+        return super().get(description)
 
 
 class ObjectsAccessor(AccessorBase):
     def __init__(self, fs):
         def match_func(frame, description):
             return frame.description == description
-        super(ObjectsAccessor, self).__init__(frames.OBJECT_FID, fs, match_func)
+        super().__init__(frames.OBJECT_FID, fs, match_func)
 
     @requireUnicode("description", "filename")
     def set(self, data, mime_type, description="", filename=""):
@@ -1572,19 +1653,18 @@ class ObjectsAccessor(AccessorBase):
 
     @requireUnicode(1)
     def remove(self, description):
-        return super(ObjectsAccessor, self).remove(description)
+        return super().remove(description)
 
     @requireUnicode(1)
     def get(self, description):
-        return super(ObjectsAccessor, self).get(description)
+        return super().get(description)
 
 
 class PrivatesAccessor(AccessorBase):
     def __init__(self, fs):
         def match_func(frame, owner_id):
             return frame.owner_id == owner_id
-        super(PrivatesAccessor, self).__init__(frames.PRIVATE_FID, fs,
-                                               match_func)
+        super().__init__(frames.PRIVATE_FID, fs, match_func)
 
     def set(self, data, owner_id):
         priv_frames = self._fs[frames.PRIVATE_FID] or []
@@ -1600,18 +1680,17 @@ class PrivatesAccessor(AccessorBase):
         return priv_frame
 
     def remove(self, owner_id):
-        return super(PrivatesAccessor, self).remove(owner_id)
+        return super().remove(owner_id)
 
     def get(self, owner_id):
-        return super(PrivatesAccessor, self).get(owner_id)
+        return super().get(owner_id)
 
 
 class UserTextsAccessor(AccessorBase):
     def __init__(self, fs):
         def match_func(frame, description):
             return frame.description == description
-        super(UserTextsAccessor, self).__init__(frames.USERTEXT_FID, fs,
-                                                match_func)
+        super().__init__(frames.USERTEXT_FID, fs, match_func)
 
     @requireUnicode(1, "description")
     def set(self, text, description=""):
@@ -1629,11 +1708,11 @@ class UserTextsAccessor(AccessorBase):
 
     @requireUnicode(1)
     def remove(self, description):
-        return super(UserTextsAccessor, self).remove(description)
+        return super().remove(description)
 
     @requireUnicode(1)
     def get(self, description):
-        return super(UserTextsAccessor, self).get(description)
+        return super().get(description)
 
     @requireUnicode(1)
     def __contains__(self, description):
@@ -1644,8 +1723,7 @@ class UniqueFileIdAccessor(AccessorBase):
     def __init__(self, fs):
         def match_func(frame, owner_id):
             return frame.owner_id == owner_id
-        super(UniqueFileIdAccessor, self).__init__(frames.UNIQUE_FILE_ID_FID,
-                                                   fs, match_func)
+        super().__init__(frames.UNIQUE_FILE_ID_FID, fs, match_func)
 
     def set(self, data, owner_id):
         data, owner_id = b(data), b(owner_id)
@@ -1666,19 +1744,18 @@ class UniqueFileIdAccessor(AccessorBase):
 
     def remove(self, owner_id):
         owner_id = b(owner_id)
-        return super(UniqueFileIdAccessor, self).remove(owner_id)
+        return super().remove(owner_id)
 
     def get(self, owner_id):
         owner_id = b(owner_id)
-        return super(UniqueFileIdAccessor, self).get(owner_id)
+        return super().get(owner_id)
 
 
 class UserUrlsAccessor(AccessorBase):
     def __init__(self, fs):
         def match_func(frame, description):
             return frame.description == description
-        super(UserUrlsAccessor, self).__init__(frames.USERURL_FID, fs,
-                                               match_func)
+        super().__init__(frames.USERURL_FID, fs, match_func)
 
     @requireUnicode("description")
     def set(self, url, description=""):
@@ -1695,19 +1772,18 @@ class UserUrlsAccessor(AccessorBase):
 
     @requireUnicode(1)
     def remove(self, description):
-        return super(UserUrlsAccessor, self).remove(description)
+        return super().remove(description)
 
     @requireUnicode(1)
     def get(self, description):
-        return super(UserUrlsAccessor, self).get(description)
+        return super().get(description)
 
 
 class PopularitiesAccessor(AccessorBase):
     def __init__(self, fs):
         def match_func(frame, email):
             return frame.email == email
-        super(PopularitiesAccessor, self).__init__(frames.POPULARITY_FID, fs,
-                                                   match_func)
+        super().__init__(frames.POPULARITY_FID, fs, match_func)
 
     def set(self, email, rating, play_count):
         flist = self._fs[frames.POPULARITY_FID] or []
@@ -1724,18 +1800,17 @@ class PopularitiesAccessor(AccessorBase):
         return popm
 
     def remove(self, email):
-        return super(PopularitiesAccessor, self).remove(email)
+        return super().remove(email)
 
     def get(self, email):
-        return super(PopularitiesAccessor, self).get(email)
+        return super().get(email)
 
 
 class ChaptersAccessor(AccessorBase):
     def __init__(self, fs):
         def match_func(frame, element_id):
             return frame.element_id == element_id
-        super(ChaptersAccessor, self).__init__(frames.CHAPTER_FID, fs,
-                                               match_func)
+        super().__init__(frames.CHAPTER_FID, fs, match_func)
 
     def set(self, element_id, times, offsets=(None, None), sub_frames=None):
         flist = self._fs[frames.CHAPTER_FID] or []
@@ -1754,10 +1829,10 @@ class ChaptersAccessor(AccessorBase):
         return chap
 
     def remove(self, element_id):
-        return super(ChaptersAccessor, self).remove(element_id)
+        return super().remove(element_id)
 
     def get(self, element_id):
-        return super(ChaptersAccessor, self).get(element_id)
+        return super().get(element_id)
 
     def __getitem__(self, elem_id):
         """Overiding the index based __getitem__ for one indexed with chapter
@@ -1772,7 +1847,7 @@ class TocAccessor(AccessorBase):
     def __init__(self, fs):
         def match_func(frame, element_id):
             return frame.element_id == element_id
-        super(TocAccessor, self).__init__(frames.TOC_FID, fs, match_func)
+        super().__init__(frames.TOC_FID, fs, match_func)
 
     def __iter__(self):
         tocs = list(self._fs[self._fid] or [])
@@ -1814,10 +1889,10 @@ class TocAccessor(AccessorBase):
         return toc
 
     def remove(self, element_id):
-        return super(TocAccessor, self).remove(element_id)
+        return super().remove(element_id)
 
     def get(self, element_id):
-        return super(TocAccessor, self).get(element_id)
+        return super().get(element_id)
 
     def __getitem__(self, elem_id):
         """Overiding the index based __getitem__ for one indexed with table
@@ -1832,7 +1907,7 @@ class TagTemplate(string.Template):
     idpattern = r'[_a-z][_a-z0-9:]*'
 
     def __init__(self, pattern, path_friendly="-", dotted_dates=False):
-        super(TagTemplate, self).__init__(pattern)
+        super().__init__(pattern)
 
         if type(path_friendly) is bool and path_friendly:
             # Previous versions used boolean values, convert old default to new